feat: finish onboarding
